requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
celery>=5.3.0
factory_boy>=3.3.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.9
Faker>=19.0.0
requests>=2.31.0
redis>=5.0.0
coverage>=7.3.2